+++ paddle/fluid/framework/CMakeLists.txt
@@ -94,8 +94,8 @@ else()
   set(parallel_executor_cuda_deps)
 endif()
 cc_library(parallel_executor SRCS parallel_executor.cc DEPS op_registry device_context scope
-        framework_proto backward glog lod_rank_table simple_threadpool scale_loss_grad_op_handle
-        fetch_op_handle ${parallel_executor_cuda_deps})
+         backward glog lod_rank_table simple_threadpool scale_loss_grad_op_handle
+        fetch_op_handle computation_op_handle ${parallel_executor_cuda_deps})
 
 cc_library(prune SRCS prune.cc DEPS framework_proto)
 cc_test(prune_test SRCS prune_test.cc DEPS op_info prune recurrent_op device_context)
